Greeks. All here, but only delta, gamma working.
Full end-to-end support in THIS code for all greeks (delta, gamma,
theta, vega, rho).  But due to QuantLib limitations only delta and
gamma are working as expected.

Delta and gamma are the only greeks supported by the QuantLib
Approximation Engines currently in use.  It seems that no such Engine
supports theta, vega or rho.

Because the available greeks are determined by the Approximation Engine
(and perhaps the Stochastic Process), and because this calculator is
letting the user choose the Engine and Process, the greek fields have
been configured to dynamically hide and show themselves based on
whether they've been provided by the current Engine/Process.